Create GuidedSensorTouchDataLogger.ino
Arduino IDE C project to get touch sensor readings from the CapacitiveSensor library and report them on the microcontroller's serial port.

The script also emits notes through a buzzer that guides the user through when and for how long to touch the sensor to ensure similar readings are taken on each run.